server/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.9.0
google-api-python-client==2.150.0
google-auth==2.35.0
google-cloud-aiplatform==1.71.0
google-generativeai==0.8.0
python-dotenv==1.0.1
